resolvido mais exercicios e alterado alguns

+++ Exercicios/Ex003/Ex003.cs
@@ -4,15 +4,15 @@ using System;
 class Exercicio3{
     static void Main(){
         Console.Write("Digite a primeira Nota: ");
-        int n1 = Convert.ToInt32(Console.ReadLine());
+        double n1 = Convert.ToInt32(Console.ReadLine());
         Console.Write("Digite a segunda Nota: ");
-        int n2 = Convert.ToInt32(Console.ReadLine());
+        double n2 = Convert.ToInt32(Console.ReadLine());
         Console.Write("Digite a terceira Nota: ");
-        int n3 = Convert.ToInt32(Console.ReadLine());
+        double n3 = Convert.ToInt32(Console.ReadLine());
         Console.Write("Digite a quarta nota: ");
-        int n4 = Convert.ToInt32(Console.ReadLine());
+        double n4 = Convert.ToInt32(Console.ReadLine());
 
-        int media = (n1 + n2 + n3 + n4) / 4;
+        double media = (n1 + n2 + n3 + n4) / 4;
         Console.Write("A soma das notas " + n1 + ", " + n2 + ", " + n3 + ", " + n4 + " Tem a media de: " + media);
     }
 }
